4.6.0対応

+++ sql/schema/view_bd001.sql
@@ -14,8 +14,8 @@ CREATE VIEW view_bd001 AS
            tbl_ptinf c 
      WHERE ((((((((a.hospnum = b.hospnum) 
        AND (a.ptid = b.ptid)) 
-       AND ((a.hospnum)::numeric = c.hospnum)) 
-       AND ((a.ptid)::numeric = c.ptid)) 
+       AND (a.hospnum = c.hospnum)) 
+       AND (a.ptid = c.ptid)) 
        AND (a.denpjtikbn <> '3'::bpchar)) 
        AND (a.denpjtikbn <> '7'::bpchar)) 
        AND (a.createkbn <> '3'::bpchar))
